Add Darwin Core mapping of events class to Rmarkdown

This function enables to easily add a code chunk with the fields which usually needs to be mapped when describing an events type for OBIS (as well as GBIF).

Usage

add_dwc_event(
  markup = "<!--DWC_EVENT-->",
  full = FALSE,
  acode = "dwc",
  object_name = "dataset",
  rmd_file = NULL
)

Arguments

markup

the identifier which will be replaced by the code chunk containing the code to perform the mapping (see details).

full

logical indicating if you want all the fields available for the event class or just the essential ones (when FALSE).

acode

a code to be added to the start of the new column names. This enables easy tracking of the new fields and easy cleanup at the end of the process.

object_name

the name of the object from which the original fields will come.

rmd_file

an optional path to an Rmd file to be edited. If NULL, then the usual Rmd file of the package is used.

Value

an edited Rmarkdown

Details

For this function to work, you need to first add the markup to the Rmarkdown. The standard one is <!--DWC_EVENT--> (with no quotation marks). Then, the function will locate the markup and substitute it for the formated code chunk. Note that you can use any markup, but this have to be identifiable within your file.

Note that the full does not contains all fields available in DwC for this class. Also, the recommended fields for the class include some other fields that should usually be included, but are not necessary. See the references for more information. Special attention should be taken when describing events and associated occurrences as some additional fields may be needed and others omitted in each table.

See also

Examples

if (FALSE) {
add_dwc_event()
}