Generate an EML metadata file in the XML format to inclusion in the IPT

EML (Ecological Markup Language) is a standard format for metadata widely used by the biodiversity databases. This function creates an standard EML file for use when submitting data through one IPT (https://ipt.gbif.org/manual/en/ipt/latest/).

Usage

gen_metadata(
  outfolder = "data/processed",
  outfile = "metadata",
  title,
  creator,
  metadataProvider,
  language,
  metadataLanguage,
  abstract,
  associatedParty = NULL,
  coverage = NULL,
  purpose = NULL,
  contact = NULL,
  methods = NULL,
  project = NULL,
  intellectualRights = NULL,
  keywordSet = NULL
)

Arguments

outfolder

the folder where the xml will be saved

outfile

the name of the metadata file

title

the title of the dataset

creator

the creator of the dataset. Should be of meta_person class, created with gen_meta_person()

metadataProvider

the provider of the dataset. Should be of meta_person class, created with gen_meta_person()

language

the language of the dataset (ISO conform)

metadataLanguage

the language of the metadata (ISO conform)

abstract

abstract of the project (character)

associatedParty

associated parties. Should be of meta_person class, created with gen_meta_person()

coverage

geographical, taxonomic or/and temporal coverage. Should be a list of objects of class meta_cov created using gen_meta_geocov(), gen_meta_temporalcov() or gen_meta_taxoncov()

purpose

the purpose of the dataset

contact

the contacts of the dataset. Should be of meta_person class, created with gen_meta_person()

methods

methods of the project. Should be a meta_methods object created with gen_meta_methods().

project

project info created with gen_meta_project()

intellectualRights

intellectual rights (text)

keywordSet

keywords. Should be a list of meta_keyword objects created with gen_meta_keywords()

Value

an XML file in the EML language

Details

Some of the fields need objects created with other functions of the "gen_meta" family. When the field asks for a list you can also supply a single object, which will then be converted into a list. For more information about the fields, consult the following link: https://www.gbif.org/sites/default/files/gbif_resource/resource-80640/gbif_metadata_profile_guide_en_v1.pdf

The & symbol should NOT be used in any text, otherwise the reading of the xml will fail.

When supplying names for the fields that require an object of type meta_person, you can select just a couple of names of the full object by using the "+" or "-" method (for more details see gen_meta_person()).

See also

Examples

if (FALSE) {
person <- gen_meta_person("John", "Doe", "Researcher", "Institute")
gen_metadata(title = "My title", creator = person, metadataProvider = person,
language = "English", metadataLanguage = "English", abstract = "Some text.")
}