Add funding details on README and .Rmd files

Add funding details on README and .Rmd files

Usage

get_figshare(
  article_id,
  download_files = FALSE,
  save_meta = FALSE,
  path = NULL
)

Arguments

article_id

the FigShare article ID. This is the number shown at the end of any FigShare document URL (e.g. https://figshare.com/articles/dataset/A_comprehensive_dataset_of_the_cold-water_coral_diversity/21997559, in this case the ID is 21997559)

download_files

logical if TRUE, then the files associated to this FigShare document will be downloaded to the "data/raw" folder. A large timeout is set, but depending on the file size it's advised to do the download separately. If any download fail, the user will receive a message.

save_meta

logical indicating if the metadata should be saved in the "data/raw" folder.

path

the path to save the files and metadata. By standard, it will save in "data/raw", but if you are working in a Rmd file, knitting will take place from the folder where the Rmd file is residing. In this case you can either supply an adjusted path or use here::here().

Value

A list with information about the FigShare article (see details).

Details

The returned list will contain the following fields:

title

authors

description (usually the abstract)

version (the version of the dataset)

date

doi

license

url

download_url (the URLs to download the files; a list with named (file name) lists containing the URLs)

All those details are relevant metadata that should ideally be included in your document. Note that some fields may be empty or with a "try-error", which indicates that the field was not available.

Examples

if (FALSE) {
get_figshare(21997559)
}